Update Ch13-multiple-lab.ipynb
I have fixed a minor error in the text description for re-sampling p-value in ch13 Jupyter file.

P-value is the fraction of the times that the test statistics obtained via re-sampling exceeds our observed test statistic (in absolute value).

It was written the other way around. The code correctly matches with the revised description.